Workaround an openbox bug of losing the focus.
I experienced a non-deterministic bug of focusing the wrong window.
It was reproducible most of the time by the window switching sequence
alt-1, alt-2, alt-0, where alt-1 and -0 switched to some window on desktop #0
while alt-2 switched to desktop #1. For details please see the state of
xbindkeysrc before this commit.

I believe this is a bug of openbox - not wmctrl or xbindkeys.
A bit of googling gave me these (hopefully relevant) links:

https://bugzilla.icculus.org/show_bug.cgi?id=3648
https://bugzilla.icculus.org/show_bug.cgi?id=4873
http://git.icculus.org/?p=dana
    /openbox.git;a=commitdiff_plain;h=197ee0ad4732c8668fe38a6b49e0a747866ef3ea
http://standards.freedesktop.org/wm-spec/wm-spec-1.3.html:w